requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
gitpython>=3.1.0
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
docling>=2.0.0
pytest>=8.0.0
